growth-database-backend/requirements.txt
astropy==5.0.4
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.12.0
healpy==1.15.2
matplotlib==3.5.2
numpy==1.22.3
pytz==2021.1
scipy==1.8.0
urllib3==1.26.5